Crate subspace_process

Structs

AsyncJoinOnDrop

Joins async join handle on drop. This future is fused, and will return Poll::Pending if polled after completion.

Functions

init_logger

raise_fd_limit

Raise the file descriptor limit for the process to the maximum possible value.

run_future_in_dedicated_thread

Runs future on a dedicated thread with the specified name. Will block on drop until background thread with future is stopped, ensuring nothing is left in memory.

set_exit_on_panic

Install a panic handler which exits on panics, rather than unwinding. Unwinding can hang the tokio runtime waiting for stuck tasks or threads.

shutdown_signal

Wait for the process to receive a shutdown signal, and log the supplied process kind.